wip: more TagSelector code cleanup
