+++ client/src/lib/modal-ids.ts
@@ -8,5 +8,5 @@ export default {
 		home: "home-notes",
 	},
 	detailedActivity: "detailed-activity", // TODO: do we want this to be dependent on activity_id?
-	tagTree: "tag-tree",
+	tagTree: { tree: "tag-tree", tagSelector: "tag-tree-tag-selector" },
 } as const;
